project1 temporarily removed camera from game which pans towards nearby text
it was taking way to much time to implement a feature which is hard to scale, and might even be worse then the current purely player centric camera that is stable and implemented now. this is because you kind of lose a sense of the player (especially considering how abstract the avatar is) when it isn't centered in the screen, also you lose some of the sense of exploration and discovery if the camera pans for you.