proj3 refactoring entire project into my ECS paradigm
Now each entity contains a mesh (pure data) which is rendered and manipulated by a system using the entities position and the global camera position to translate the thing, and the camera rotation and the entities angular velocity to rotate the thing appropriately.

hopefully using this approach I will be able to have an arbitrary number of meshes and will be able to have a movable, rotatable camera.